Update doctrine-lifecycle-callback.rst
The point of this code is that you have to enable it using the PrePersist annotation.  I would assume inadvertently that the code was changed in the prePersist class method to set the class variable directly rather than call the setCreatedAt() method in the first example block.  I just fixed this for consistency and to remove the confusion.